Makefile finished? Libft funcs shored up so it actually compiles

+++ libft/ft_putllong_base_fd.c
@@ -6,16 +6,16 @@
 /*   By: jmeier <jmeier@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/09/20 12:52:12 by jmeier            #+#    #+#             */
-/*   Updated: 2017/11/28 14:43:16 by jmeier           ###   ########.fr       */
+/*   Updated: 2017/11/29 11:50:35 by jmeier           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_putllong_base_fd(long long n, int base, int fd)
+void	ft_putllong_base_fd(long long n, long long base, int fd)
 {
-	int		p;
-	char	*c;
+	long long	p;
+	char		*c;
 
 	p = 1;
 	c = "0123456789ABCDEFHIJKLMNOPQRSTUVWXYZ";
